pyproject.toml
[tool.poetry]
name = "sirius"
version = "0.1.0"
description = ""
package-mode = false
authors = [""]

[tool.poetry.dependencies]
aiohttp = "3.9.5"
asyncpg = "0.29.0"
orjson = "3.10.6"
pydantic = "2.8.2"
pydantic-settings = "2.3.4"
python = "^3.11"
pyyaml = "^6.0"
sqlalchemy = "2.0.31"
alembic = "1.13.2"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
pyjwt = "^2.9.0"
greenlet = "^3.0.3"
uvloop = "^0.20.0"
aiogram = "^3.13.0"
redis = "^5.0.8"
hiredis = "^3.0.0"
aio-pika = "^9.4.3"
msgpack = "^1.1.0"
starlette-context = "^0.3.6"
jinja2 = "^3.1.4"
msgpack-types = "^0.5.0"
prometheus-client = "^0.21.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
asgi-lifespan = "^2.1.0"

[tool.poetry.group.dev.dependencies]
autoflake = "1.4"
black = "24.4.2"
flake8 = "7.1.0"
flake8-black = "0.3.6"
flake8-bugbear = "24.4.26"
flake8-coding = "1.3.2"
flake8-comprehensions = "3.15.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.5.0"
flake8-executable = "2.1.3"
flake8-isort = "6.1.1"
flake8-logging = "1.6.0"
flake8-pie = "0.16.0"
flake8-pytest-style = "2.0.0"
flake8-quotes = "3.4.0"
flake8-simplify = "0.21.0"
isort = "5.13.2"
mypy = "1.11.0"
types-passlib = "^1.7.7.10"
types-python-jose = "^3.3.4.5"
types-pyyaml = "^6.0.12.8"

[tool.black]
line-length = 120
target-version = ["py312"]
skip-string-normalization = true
extend-exclude = '''
^/(
  (
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | venv
  )/
)
'''

[tool.pytest.ini]
env = [
    "DB_URL=....",
    "DB_URL=....",
    "DB_URL=....",
    "DB_URL=....",
]

[tool.isort]
line_length = 120
combine_as_imports = true
order_by_type = true
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3

[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy",
]
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_reexport = false
exclude = [
    "^conf/",
    "^tests/",
    "^utils/",
    "^.venv/",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = false  # to allow parsing
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = [
    "asyncpg.*",
    "prometheus_client.*",
    "starlette_prometheus.*",
    "uvicorn.*",
]
ignore_missing_imports = true

